// src/app/core/interfaces/task.ts
export interface Task {
  id: string;
  name: string;
  time: string; 
  isActive: boolean;
  isCompleted: boolean;
  //  name: string;
  // date: string;  // بدل time عشان نستخدم تاريخ ووقت كامل بصيغة ISO
  // isDateAndTimeEnded: boolean;
  // isCompleted: boolean;
}